+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.0.4] - 2023-06-09
+
 ### Added
 
 - `Client.download_item_collection()` ([#20](https://github.com/gadomski/stac-asset/pull/20))
@@ -51,7 +53,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Initial release.
 
-[unreleased]: https://github.com/gadomski/stac-asset/compare/v0.0.3...HEAD
+[unreleased]: https://github.com/gadomski/stac-asset/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/gadomski/stac-asset/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/gadomski/stac-asset/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/gadomski/stac-asset/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/gadomski/stac-asset/releases/tag/v0.0.1